Move FoldedModelEntity to parent package.
