+++ src/main/java/oripa/persistence/entity/exporter/FoldedModelExporterSVG.java
@@ -30,6 +30,7 @@ import java.util.List;
 import oripa.domain.fold.halfedge.OriFace;
 import oripa.domain.fold.halfedge.OrigamiModel;
 import oripa.domain.fold.origeom.OverlapRelation;
+import oripa.persistence.entity.FoldedModelEntity;
 import oripa.persistence.filetool.Exporter;
 import oripa.persistence.svg.FacesToSvgConverter;
